source/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.3
Django==4.1.7
django-address==0.2.8
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.9.0
django-angularjs==1.0.2
django-background-task==0.1.8
django-bootstrap-modal-forms==3.0.4
django-bootstrap-v5==1.0.11
django-bootstrap4==22.3
django-bootstrap5==23.3
django-cors-headers==4.2.0
django-daterange-filterspec==2.0.4
django-encrypted-secrets==0.9.9
django-forms-bootstrap==3.1.0
django-mathfilters==1.0.0
django-polymorphic==3.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
medisearch-client==0.2.0
multidict==6.0.4
mysqlclient==2.1.1
names==0.3.0
ndg-httpsclient==0.5.1
Pillow==9.4.0
psycopg2==2.9.5
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
PyMySQL==1.0.3
pyOpenSSL==23.2.0
pyral==1.5.2
python-dotenv==1.0.0
python-graphql-client==0.4.3
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
websocket-client==1.6.2
websockets==11.0.3
yarl==1.9.2
